Changing TreeSet to HashMap
This changes the TreeSet to a HashMap. A (mathematical) set is a set of
distinct objects. A TreeSet's compare function determines relation of
two objects. It also determines equality and since this is a Set it will
effectively remove one of the two objects from the set. Although the two
objects merely share the same priority and not the same origin or
fromAddress. This will lead to an inconsistency in the representation of
presence objects and resources that contact has connected. The resource
not represented in this TreeSet would no longer be able to let the
contact appear as Online.